requirements.txt
tensorflow>=2.10.0
numpy>=1.21.0
Pillow>=9.0.0